adds restaurant and navbar improvements
