--- src/app/home/restaurants/restaurants.module.css
@@ -1,27 +1,0 @@
-.cardContainer {
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  gap: 20px;
-  padding: 20px;
-}
-
-.card {
-  width: 200px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  overflow: hidden;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.cardImage {
-  width: 100%;
-  height: 150px;
-  object-fit: cover;
-}
-
-.cardName {
-  padding: 10px;
-  text-align: center;
-  font-weight: bold;
-}
