+++ src/app/navigation/navbar.css
@@ -1,12 +1,32 @@
 .navbar {
-  background-color: #333;
-  padding: 1rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #4caf50;
+  padding: 10px 20px;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+}
+
+.logoImage {
+  width: 40px;
+  height: 40px;
+  margin-right: 10px;
+}
+
+.logoName {
+  font-size: 1.5rem;
+  font-weight: bold;
+  color: white;
 }
 
 .navList {
-  list-style: none;
   display: flex;
   justify-content: space-around;
+  list-style: none;
   margin: 0;
   padding: 0;
 }
@@ -14,7 +34,9 @@
 .navItem a {
   color: white;
   text-decoration: none;
-  font-size: 1.2rem;
+  font-size: 1rem;
+  font-weight: bold;
+  margin-left: 20px;
 }
 
 .navItem a:hover {
