+++ src/app/navigation/navbar.jsx
@@ -4,30 +4,28 @@ import './navbar.css';
 const Navbar = () => {
   return (
     <nav className="navbar">
+      <div className="logo">
+        <img src="/images/logo.png" alt="Foodcourt Logo" className="logoImage" />
+        <span className="logoName">Foodcourt</span>
+      </div>
       <ul className="navList">
         <li className="navItem">
-          <Link href="/login">Login</Link>
-        </li>
-        <li className="navItem">
           <Link href="/home">Home</Link>
-        </li>
-        <li className="navItem">
-          <Link href="/home/restaurants">Restaurants</Link>
         </li>
         <li className="navItem">
           <Link href="/menu">Menu</Link>
         </li>
         <li className="navItem">
-          <Link href="/tables">Book Tables</Link>
+          <Link href="/tables">Tables</Link>
         </li>
         <li className="navItem">
           <Link href="/checkout">Checkout</Link>
         </li>
         <li className="navItem">
-          <Link href="/waiting">Waiting Time</Link>
+          <Link href="/dashboard">Dashboard</Link>
         </li>
         <li className="navItem">
-          <Link href="/dashboard">Dashboard</Link>
+          <Link href="/logout">Logout</Link>
         </li>
       </ul>
     </nav>
